Modified CLI access cuz i wasn't able to do it before, now it works just with 'cargo run'. Also, implemented stuff done for voter and voting process. Currently 3 voters in the database, 1 is tested, i've set it in a way so that a voter can only vote once in an election. Voter 2 is Spongebob (pin 1124) and Patrick (pin 1125), use these two to test the voting part for me. Oh and for some reason, all of the info related to elections, voters, admin are being saved to rusttrust.db (root folder) instead of elections.db (data folder), I don't know why.
Last things to note: There is no check in place to login as an Admin to do admin operations. I think this should be fixed. Additionally, It would be nice to implement an option to view all the voters as a list. There's some dead code in admin.rs since some of the functions aren't being used yet.